Update Paul's Search Page 2.0.html
OK I've just updated this as raw code- Had to review the naming mechanisms. 

The first thing is the "anchor": "#item" in the nav link, then that name has to be used in the id element inside the tab content.

The id in the nav link must not be the same as the name or else the ref link will malfunction. It's easy to mix these two things up. e.g. the reference that says "home-tab" has no relationship to the href "#cats" which is mirrored inside the tab content as id="cats".